Support DWRF stripes above MAX_INT rows in reader
OrcReader fails when a stripe has more than INT32_MAX rows.
Both ORC/DWRF format uses long for stripe number of rows.
We ran into this edge case on a file with only string as a column
and file is produced by a different writer compatible with
DWRF spec.

+++ presto-orc/src/main/java/com/facebook/presto/orc/OrcWriterStats.java
@@ -45,7 +45,7 @@ public class OrcWriterStats
             StripeInformation stripeInformation)
     {
         long stripeBytes = stripeInformation.getTotalLength();
-        int stripeRows = stripeInformation.getNumberOfRows();
+        long stripeRows = stripeInformation.getNumberOfRows();
         getFlushStats(flushReason).recordStripeWritten(stripeBytes, stripeRows, dictionaryBytes);
         allFlush.recordStripeWritten(stripeBytes, stripeRows, dictionaryBytes);
     }
